requirements.txt
alembic==1.13.1
async-timeout==4.0.3
audioread==3.0.1
blinker==1.7.0
certifi==2023.11.17
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
decorator==5.1.1
docopt==0.6.2
Flask==3.0.0
Flask-Cors==4.0.0
Flask-Migrate==4.0.7
Flask-SQLAlchemy==3.1.1
idna==3.6
itsdangerous==2.1.2
Jinja2==3.1.2
joblib==1.3.2
lazy_loader==0.3
librosa==0.10.1
llvmlite==0.41.1
Mako==1.3.3
MarkupSafe==2.1.3
msgpack==1.0.7
numba==0.58.1
numpy==1.26.2
packaging==23.2
platformdirs==4.1.0
pooch==1.8.0
psutil==5.9.7
psycopg2==2.9.9
pycparser==2.21
pydub==0.25.1
pyee==11.1.0
python-ffmpeg==2.0.11
redis==4.5.4
requests==2.31.0
rq==1.10.0
scikit-learn==1.3.2
scipy==1.11.4
soundfile==0.12.1
soxr==0.3.7
SQLAlchemy==2.0.29
threadpoolctl==3.2.0
typing_extensions==4.9.0
urllib3==2.1.0
ultralytics==8.0.200
Werkzeug==3.0.1
opencv-python==4.10.0.84
